requirements.txt
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database: built-in sqlite3 (no extra deps, Python 3.14 compatible)

# AI & APIs
anthropic==0.18.1
google-auth==2.27.0
google-auth-oauthlib>=1.2.0,<2
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Auth
pyjwt>=2.8.0
cryptography>=42.0.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.26.0
jinja2==3.1.3
